remove local SSWU implementation and rely only on BLST map to G1 tools
